backend_v2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.26.0
Pillow>=10.3.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.0.2
mediapipe==0.10.7
opencv-python==4.8.1.78
httpx>=0.24.0,<0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
twilio==8.10.0
